requirements.txt
# ===== Core Dependencies =====
blinker==1.9.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1

# ===== Flask & Extensions =====
Flask==3.1.1
flask-cors==6.0.1
Werkzeug==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2

# ===== WSGI Server =====
gunicorn==23.0.0

# ✅ gevent 제거 - gthread 워커 사용으로 불필요
# gevent는 ONNX Runtime과 호환성 이슈 발생 (segfault)
# gthread (스레드 기반)가 더 안정적이고 메모리 효율적

# ===== ONNX Runtime =====
# ✅ 1.19.2 버전 유지 - Render Free tier 안정성 검증됨
# 1.22.x 이상에서 segfault 및 메모리 이슈 발생
onnxruntime==1.19.2
onnx==1.17.0
protobuf==3.20.3

# ===== PyTorch (ResNet50 Upgrade) =====
# TensorFlow 제거(미사용) 및 PyTorch로 전환
# Windows/CPU 환경에서 기본 PyPI 빌드 사용
torch
torchvision

# ===== Image Processing =====
pillow==11.0.0
numpy==1.26.4  # onnxruntime 1.19.2 호환

# ===== System Monitoring =====
psutil==5.9.8
humanfriendly==10.0

# ===== Observability - Phase 4 =====
# Prometheus 메트릭 export (production-grade monitoring)
prometheus-client==0.20.0

# ===== Environment Management =====
python-dotenv==1.0.0

# ===== Utilities =====
requests==2.32.4
urllib3==2.5.0
idna==3.10
six==1.17.0

# ===== Math Libraries =====
mpmath==1.3.0
sympy==1.14.0

# ===== Build Tools =====
# ✅ Render 배포 안정성 향상을 위한 빌드 도구
setuptools==75.8.0
wheel==0.45.1
packaging==25.0

# ===== Optional Dependencies =====
flatbuffers==25.2.10

# ===== Testing Dependencies =====
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
